Changed initialisation
